prevent NPE in dispose()
if the editor was not fully initialized NPEs can be thrown in dispose().
e.g. when the editor is used as page in a multipage editor and the page
was never selected.

+++ plugins/org.eclipse.bpmn2.modeler.ui/src/org/eclipse/bpmn2/modeler/ui/editor/BPMN2Editor.java
@@ -735,7 +735,9 @@ public class BPMN2Editor extends DiagramEditor implements IPreferenceChangeListe
 
 	@Override
 	public void dispose() {
-		getTargetRuntime().notify(new LifecycleEvent(EventType.EDITOR_SHUTDOWN,this));
+		if (targetRuntime != null) {
+			targetRuntime.notify(new LifecycleEvent(EventType.EDITOR_SHUTDOWN,this));
+		}
 
 		// clear ID mapping tables if no more instances of editor are active
 		int instances = 0;
@@ -750,7 +752,7 @@ public class BPMN2Editor extends DiagramEditor implements IPreferenceChangeListe
 			}
 		}
 		BPMN2Editor otherEditor = findOpenEditor(this, getEditorInput());
-		if (otherEditor==null) {
+		if (otherEditor==null && diagramUri != null) {
 			// we can delete the Graphiti Diagram file if there are no other
 			// editor windows open for this BPMN2 file.
 			File diagramFile = new File(diagramUri.toFileString());
@@ -766,7 +768,9 @@ public class BPMN2Editor extends DiagramEditor implements IPreferenceChangeListe
 			ModelUtil.clearIDs(modelHandler.getResource(), instances==0);
 			modelHandler.dispose();
 		}
-		getPreferences().removePreferenceChangeListener(this);
+		if (preferences != null) {
+			preferences.removePreferenceChangeListener(this);
+		}
 		
 		// cancel the Property Sheet Page job
 		if (propertySheetPage!=null)
@@ -776,24 +780,32 @@ public class BPMN2Editor extends DiagramEditor implements IPreferenceChangeListe
 		if (tabDescriptorProvider instanceof PropertyTabDescriptorProvider)
 			((PropertyTabDescriptorProvider)tabDescriptorProvider).disposeTabDescriptors(bpmnResource);
 		
-		getResourceSet().eAdapters().remove(getEditorAdapter());
+		if (getResourceSet() != null) {
+			getResourceSet().eAdapters().remove(getEditorAdapter());
+		}
 		removeSelectionListener();
 		if (instances==0)
 			setActiveEditor(null);
 		
 		super.dispose();
-		ModelHandlerLocator.remove(modelUri);
-		// get rid of temp files and folders, but NOT if the workbench is being shut down.
-		// when the workbench is restarted, we need to have those temp files around!
-		if (!workbenchShutdown) {
-			if (FileUtils.isTempFile(modelUri)) {
-				FileUtils.deleteTempFile(modelUri);
+		if (modelUri != null) {
+			ModelHandlerLocator.remove(modelUri);
+			// get rid of temp files and folders, but NOT if the workbench is being shut down.
+			// when the workbench is restarted, we need to have those temp files around!
+			if (!workbenchShutdown) {
+				if (FileUtils.isTempFile(modelUri)) {
+					FileUtils.deleteTempFile(modelUri);
+				}
 			}
 		}
 
 		removeWorkbenchListener();
 		removeMarkerChangeListener();
-		getPreferences().dispose();
+
+		if (preferences != null) {
+			preferences.dispose();
+		}
+
 		currentInput = null;
 	}
